+++ test/unit/git/test_autosubmit_git.py
@@ -16,13 +16,8 @@
 # along with Autosubmit.  If not, see <http://www.gnu.org/licenses/>.
 
 """Tests for ``AutosubmitGit``."""
-from getpass import getuser
-from pathlib import Path
-from typing import Callable
 
 import pytest
-from pytest_mock import MockerFixture
-from ruamel.yaml import YAML
 
 from autosubmit.git.autosubmit_git import AutosubmitGit
 from log.log import AutosubmitCritical
@@ -198,61 +193,3 @@ def test_copy_code(autosubmit_config, config, mocker, autosubmit):
 )
 def test_valid_git_repo_check(git_repo: str, expected: str) -> None:
     assert AutosubmitGit.is_git_repo(git_repo) == expected
-
-
-def _get_experiment_data(tmp_path):
-    _user = getuser()
-
-    return {
-        'PLATFORMS': {
-            'pytest-ps': {
-                'type': 'ps',
-                'host': '127.0.0.1',
-                'user': _user,
-                'project': 'whatever',
-                'scratch': str(tmp_path / 'scratch'),
-                'DISABLE_RECOVERY_THREADS': 'True'
-            }
-        },
-        'JOBS': {
-            'debug': {
-                'SCRIPT': 'echo "Hello world"',
-                'RUNNING': 'once'
-            },
-        }
-    }
-
-
-@pytest.mark.parametrize(
-    "project_type,expid",
-    [
-        ('git', 'o001'),
-        ('git', 'a001'),
-        ('git', 't001'),
-        ('git', 'e001'),
-    ]
-)
-def test_remote_repo_operational(project_type: str, expid: str, tmp_path: Path, autosubmit_exp: Callable,
-                                 autosubmit, mocker: MockerFixture) -> None:
-    """
-    Tests the check_unpushed_changed function from AutosubmitGit.
-
-    Ensures no operational test with unpushed changes in their Git repository is run.
-    """
-    as_exp = autosubmit_exp(expid, _get_experiment_data(tmp_path))
-    run_dir = Path(as_exp.as_conf.basic_config.LOCAL_ROOT_DIR)
-    temp_path = run_dir / expid / "conf" / f"expdef_{expid}.yml"
-
-    with open(temp_path, 'r') as f:
-        yaml = YAML(typ='rt')
-        data = yaml.load(f)
-    data["PROJECT"]["PROJECT_TYPE"] = project_type
-    with open(temp_path, 'w') as f:
-        yaml.dump(data, f)
-
-    mocker.patch('subprocess.check_output', return_value=b'M\n')
-    if expid[0] != 'o':
-        AutosubmitGit.check_unpushed_changes(expid)
-    else:
-        with pytest.raises(AutosubmitCritical):
-            AutosubmitGit.check_unpushed_changes(expid)